requirements.txt
attrs~=25.4.0
typing_extensions~=4.15.0
SQLAlchemy~=2.0.44
greenlet~=3.2.4
jinja2~=3.1.6
python-vlc~=3.0.21203
PyQt5~=5.15.11
PyQtWebEngine~=5.15.7
Pillow>=12.0.0
pyinstaller~=6.17.0
pydantic~=2.12.5
requests>=2.32.3
aiohttp>=3.13.2
httpx~=0.28.1
beautifulsoup4~=4.14.3
python-vlc~=3.0.21203
python-mpv~=1.0.8

# Not using now
pytest~=8.3.5
python-dotenv~=1.1.0
qrcode~=8.2

# Not needed for last build
# requests_toolbelt~=1.0.0
# tqdm~=4.67.1
# pyzbar~=0.1.9
# pyzipper~=0.3.6
